Use workqueue to perform cleanup at process exit
Turns out, it's not allowed to sleep while a process is exiting.
Our do_exit kprobe handler could do so when acquiring a write lock to
the client info list, resulting in a "scheduling while atomic" BUG in
dmesg.

This commit fixes that by changing the handler to just schedule the
cleanup for later.